Polish the code for panels
- Remove dead code
- Split controls code into multiple files for readibility
- Organize imports
- Add a stub for ComboBox
