+++ src/app/views/panels/index.tsx
@@ -1,2 +1,1 @@
-export * from "./attribute_editor";
 export * from "./attribute_panel";
