+++ src/app/views/panels/widgets/controls/index.tsx
@@ -1,0 +1,10 @@
+export * from "./button";
+export * from "./combo_box";
+export * from "./image";
+export * from "./input_color";
+export * from "./input_expression";
+export * from "./input_file";
+export * from "./input_number";
+export * from "./input_text";
+export * from "./select";
+export * from "./slider";
